+++ tests/data/importer/simple_pkg/a.py
@@ -1,5 +1,0 @@
-import simple_pkg.b
-
-
-def subtract_one_add_two(n):
-    return simple_pkg.b.subtract_one(n) + 2