+++ tests/fork_test.py
@@ -24,7 +24,6 @@ import mitogen.core
 import mitogen.fork
 
 import testlib
-import plain_old_module
 
 
 def _find_ssl_linux():
@@ -90,9 +89,8 @@ def exercise_importer(n):
     """
     Ensure the forked child has a sensible importer.
     """
-    sys.path.remove(testlib.DATA_DIR)
-    import simple_pkg.a
-    return simple_pkg.a.subtract_one_add_two(n)
+    import testmods.simple_pkg.a
+    return testmods.simple_pkg.a.subtract_one_add_two(n)
 
 
 skipIfUnsupported = unittest.skipIf(