+++ docs/changelog.rst
@@ -37,6 +37,7 @@ In progress (unreleased)
   as test requirement
 * :gh:issue:`1455` tests: Fix missing and unused imports of Mitogen
 * :gh:issue:`1455` CI: Remove sys.path manipulation in .ci/unit_tests.py
+* :gh:issue:`1455` tests: Consolidate modules used by tests in tests/testmods
 
 
 v0.3.42 (2026-02-20)